# docs/swagger.yaml
basePath: /api
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a learning management system API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Labpro API
  version: "1.0"
paths:
...
  /modules/{courseId}:
    get:
      description: Get a paginated list of modules for a specific course
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 15, max: 50)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              message:
                type: string
              pagination:
                type: object
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get modules of a course
      tags:
      - modules
    post:
      consumes:
      - multipart/form-data
      description: Create a new module for a specific course with title, description,
        PDF and video files
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      - description: Module title
        in: formData
        name: title
        required: true
        type: string
      - description: Module description
        in: formData
        name: description
        type: string
      - description: PDF file
        in: formData
        name: pdf_file
        type: file
      - description: Video file
        in: formData
        name: video_file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new module (Admin only)
      tags:
      - admin-modules
  /modules/{id}:
    delete:
      description: Delete an existing module and all its associated data
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a module (Admin only)
      tags:
      - admin-modules
  /modules/{id}/complete:
    post:
      description: Mark a specific module as completed by the user
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark module as completed
      tags:
      - modules
  /modules/detail/{id}:
    get:
      description: Get detailed information about a specific module
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get module by ID
      tags:
      - modules
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"